Change unset to use a hierarchical key instead of flags
This makes 'unset' a lot easier to use, but means it is no longer
symmetrical with 'set'.  If set was to also use the hierarchy then it
would be very difficult to set values as you would need to know how
the settings are arranged before hand, and if you set values that the
program doesn't understand it could all get very confusing.  With
'unset' on the other hand, you already know which values have been set
thanks to 'config' and you can easily see the hierarchy.  Maybe there
will be support for hierarchical keys in 'set' in the future, but all
it's implications will have to be figured out first.